Update create_share_links.sh
Just a change added which makes rebuilds easier. Otherwise, the resource group must be torn down and rebuilt.

+++ examples/reveal_E2E/scripts/create_share_links.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 
+if [ ! -L /share/apps ]
+then
 ln -s /apps /share/apps
 ln -s /data /share/data
+fi
